requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
mysql-connector-python==8.1.0

deepface==0.0.90
opencv-python==4.8.1.78
numpy==1.26.0
matplotlib==3.8.1
Pillow==10.0.1
tqdm==4.67.0
# pickle5==0.0.11

easyocr==1.7
rapidfuzz==3.14.3


